requirements.txt
channels==3.0.4
Django==4.0.2
django_cors_headers==3.11.0
django_extensions==3.1.5
django_filter==21.1
django_redis==5.2.0
djangochannelsrestframework==1.0.0
djangorestframework==3.13.1
drf_extensions==0.7.1
drf_writable_nested==0.6.3
mysqlclient==2.1.0
requests==2.27.1
